[CMake] Officially enroll support for Apple Clang
Apple Clang is supported now, only if the binutils are installed. This is because Apple's Clang fails when trying to link the program (because it can't find LLD). This is solved either by adding ld.lld to the PATH, by passing it to the -fuse-ld switch or by pointing the -B flag to the installed LLVM/binutils tools.

Signed-off-by: Oever González <notengobattery@gmail.com>