[CMake][x86] Fix Clang target for x86
Currently, LLVM won't link the project if the target is 'i686-elf', which is the "bare metal" toolchain. Clang will build for that target but will only link for 'i686-linux-elf' and 'i686-linux-gnu', which are the Linux Kernel targets. Unless Clang will allow the linker to run with 'i686-elf', the requiered ad-hoc flag will be included for all x86 Clang linker commands.

Signed-off-by: Oever González <notengobattery@gmail.com>