[All] Fix the linker script to keep the stack at the end
Some embedded devices, such as ARM boards, will load the image at a certain address and will begin the execution at that address. This way, it makes no sense to keep the stack in the start of the image. By letting the stack in the image, we can guarantee that all the memory is available. The stack is just after the rodata section, so it will be protected (in theory) to overflows because the MMU will refuse to write in the rodata section. Also, the data section is between the text and rodata, so it's protected by the MMU too.

Signed-off-by: Oever González <notengobattery@gmail.com>