[Python] Create a proof-of-concept script to compile static configurations
This script is intended to generate a C source code file which contains a hash table that will be populated with static configurations, which may need to be accessed quickly and easily during boot. It's a idea similar to Linux's device tree.

The input is a JSON file full of properties and the output is a hash table and a series of char* strings. The table will containt the addresses of the strings at the index represented by the hash.

Signed-off-by: Oever González <notengobattery@gmail.com>