[All] Partially implement the hash code for the Device Descriptor
This will add the Hash1, which is the first hash calculated. The program must have to verify the property against the requested property, perform a small lookup if it does not match, calculate the Hash2 if it fails and finaly perform a last lookup if it fails. From that point, the property is not in the table.

Signed-off-by: Oever González <notengobattery@gmail.com>