[All] Change 'CoreLibraries' to static and disable compiler features
Some compiler features are not valid for a standard library and/or the compiler support library. Disable them.

Also, this commit changes the CoreLibraries to be a static library, instead of an interface library. If this change is not present, the actual core libraries won't be added correctly to the linker command.

Signed-off-by: Oever González <notengobattery@gmail.com>