[All] Update FANCY_LIST feature disables
Generally speaking, LTO helps a lot by unintentionally resolving the
"dependencies", so you don't really need to disable the fancy features
in all of the libraries, just on the files that are meant to be built
that way. That is because the optimizer will recognize the flags during
linking and emit code with the updated flags (overriding the previous
flags).

Signed-off-by: Oever González <notengobattery@gmail.com>