[All] Update Device Descriptor to support custom tags
With this update, Device Descriptor can support custom tags that have a
custom output format and a custom input format. Hand in hand with it,
this allows the include logic to be improved by using tags to set the
kind of include that is being made. This allows, also, type checking and
better error reporting.

Signed-off-by: Oever González <notengobattery@gmail.com>